Use pathPattern
