+++ inlang/source-code/plugins/icu1-importer/src/settings.ts
@@ -2,11 +2,11 @@ import { Type, type Static } from "@sinclair/typebox"
 
 export type PluginSettings = Static<typeof PluginSettings>
 export const PluginSettings = Type.Object({
-	messagesFolderPath: Type.String({
-		pattern: ".*",
-		examples: ["./messages", "./i18n"],
-		title: "Path to folder containing language files",
+	pathPattern: Type.String({
+		pattern: ".*\\{locale\\}.*\\.json$",
+		examples: ["./messages/{locale}.json", "./i18n/{locale}.json"],
+		title: "Path to language files",
 		description:
-			"The relative path to the folder containing the language files in your repository. It must contain files named `{locale}.json`.",
+			"Specify the pathPattern to locate language files in your repository. It must include `{locale}` and end with `.json`.",
 	}),
 })
